moved functions from cli script into new module client.py
